third_party/blink/web_tests/external/wpt/app-history/navigate/goTo-detach-in-onnavigate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(async () => {
    let iframe_constructor = i.contentWindow.DOMException;
    let i_win = i.contentWindow;
    i.contentWindow.history.pushState(1, "", "#1");
    assert_equals(i.contentWindow.appHistory.entries().length, 2);
    let key = i.contentWindow.appHistory.entries()[0].key;

    i.contentWindow.appHistory.onnavigate = () => i.remove();
    await promise_rejects_dom(t, 'AbortError', iframe_constructor, i.contentWindow.appHistory.goTo(key));
    t.done();
  });
}, "goTo() in a detached window");
</script>
